Simplify query building on edit
`r_norm`, `r_fuzzy` and `r_to_fuzzy` are used to build lists of columns
for translations when adding a new row on modification of the original
text. They are built as `column_1, column_2..., column_n`. As they are
all used after other columns in the query, they are tested for content
before adding them to the query prepending a needed comma. Code is much
simpler if we just have that comma since the beginning.