Fix: keep fuzzyness on source text modification
When changing a translatable block in the original source, translations
are usually marked fuzzy for revision. That's not really needed for
some kind of changes, and there's an option to mark the change as not
invalidating translations. This option is not copying the fuzzy bit
from current translations, though, with the end result that a fuzzy
translation of a block changed with new text magically becomes finished.